TIDIJ 180 Move VehicleState class to a separate file
